Split get-timing-stats into separate modules
Create a timing-data module that defines:
The type TimeSet, which can be used to represent a set of timing tests
The function read_one_test(), which reads testing data into a TimeSet.

Create a program-output module that defines:
A printing function to print the contents of a TimeSet, in a nice format.
Two copy functions to copy the contents of one file into another.  One copies
until a sentinel, the other copies until the EOF.